Fix errors in BattleScene
