+++ src/Scene/BattleScene.js
@@ -29,13 +29,13 @@ const BattleScene = new Phaser.Class({
     const dragonblue = new Enemy(this, 50, 50, 'dragonblue', null, 'Dragon', 50, 3);
     this.add.existing(dragonblue);
 
-    const dragonOrange = new Enemy(this, 50, 100, 'dragonorrange', null,'Dragon2', 50, 3);
+    const dragonOrange = new Enemy(this, 50, 100, 'dragonorrange', null, 'Dragon2', 50, 3);
     this.add.existing(dragonOrange);
 
     // array with heroes
-    this.heroes = [ warrior, mage ];
+    this.heroes = [warrior, mage];
     // array with enemies
-    this.enemies = [ dragonblue, dragonOrange ];
+    this.enemies = [dragonblue, dragonOrange];
     // array with both parties, who will attack
     this.units = this.heroes.concat(this.enemies);
 
@@ -67,9 +67,9 @@ const BattleScene = new Phaser.Class({
       let r;
       do {
         r = Math.floor(Math.random() * this.heroes.length);
-      } while (!this.heroes[r].living) 
-      // call the enemy's attack function 
-      this.units[this.index].attack(this.heroes[r]);  
+      } while (!this.heroes[r].living);
+      // call the enemy's attack function
+      this.units[this.index].attack(this.heroes[r]);
       // add timer for the next turn, so will have smooth gameplay
       this.time.addEvent({ delay: 3000, callback: this.nextTurn, callbackScope: this });
     }
@@ -99,7 +99,7 @@ const BattleScene = new Phaser.Class({
       this.units[this.index].attack(this.enemies[target]);
     }
     // next turn in 3 seconds
-    this.time.addEvent({ delay: 3000, callback: this.nextTurn, callbackScope: this });        
+    this.time.addEvent({ delay: 3000, callback: this.nextTurn, callbackScope: this });
   },
 
   endBattle() {
@@ -125,7 +125,7 @@ const Unit = new Phaser.Class({
   initialize:
 
     function Unit(scene, x, y, texture, frame, type, hp, damage) {
-      Phaser.GameObjects.Sprite.call(this, scene, x, y, texture, frame)
+      Phaser.GameObjects.Sprite.call(this, scene, x, y, texture, frame);
       this.type = type;
       this.hp = hp;
       this.maxHp = this.hp;
@@ -186,7 +186,7 @@ const MenuItem = new Phaser.Class({
   initialize:
 
   function MenuItem(x, y, text, scene) {
-    Phaser.GameObjects.Text.call(this, scene, x, y, text, { color: '#ffffff', align: 'left', fontSize: 15});
+    Phaser.GameObjects.Text.call(this, scene, x, y, text, { color: '#ffffff', align: 'left', fontSize: 15 });
   },
 
   select() {
@@ -286,7 +286,7 @@ const Menu = new Phaser.Class({
   // recreate the menu items
   remap(units) {
     this.clear();
-    for (let i = 0; i < units.length; i += 1) {
+    for (let i = 0; i < units; i += 1) {
       const unit = units[i];
       unit.setMenuItem(this.addMenuItem(unit.type));
     }
@@ -334,4 +334,162 @@ const EnemiesMenu = new Phaser.Class({
   },
 });
 
-export default BattleScene;
+const UIScene = new Phaser.Class({
+
+  Extends: Phaser.Scene,
+
+  initialize:
+
+    function UIScene() {
+      Phaser.Scene.call(this, { key: 'UIScene' });
+    },
+
+  create() {
+    // draw some background for the menu
+    this.graphics = this.add.graphics();
+    this.graphics.lineStyle(1, 0xffffff);
+    this.graphics.fillStyle(0x031f4c, 1);
+    this.graphics.strokeRect(2, 150, 90, 100);
+    this.graphics.fillRect(2, 150, 90, 100);
+    this.graphics.strokeRect(95, 150, 90, 100);
+    this.graphics.fillRect(95, 150, 90, 100);
+    this.graphics.strokeRect(188, 150, 130, 100);
+    this.graphics.fillRect(188, 150, 130, 100);
+
+    // basic container to hold all menus
+    this.menus = this.add.container();
+
+    this.heroesMenu = new HeroesMenu(195, 153, this);
+    this.actionsMenu = new ActionsMenu(100, 153, this);
+    this.enemiesMenu = new EnemiesMenu(8, 153, this);
+
+    // the currently selected menu
+    this.currentMenu = this.actionsMenu;
+
+    // add menus to the container
+    this.menus.add(this.heroesMenu);
+    this.menus.add(this.actionsMenu);
+    this.menus.add(this.enemiesMenu);
+
+    this.battleScene = this.scene.get('BattleScene');
+
+    // listen for keyboard events
+    this.input.keyboard.on('keydown', this.onKeyInput, this);
+
+    // when its player cunit turn to move
+    this.battleScene.events.on('PlayerSelect', this.onPlayerSelect, this);
+
+    // when the action on the menu is selected
+    // for now we have only one action so we dont send and action id
+    this.events.on('SelectedAction', this.onSelectedAction, this);
+
+    // an enemy is selected
+    this.events.on('Enemy', this.onEnemy, this);
+
+    // when the scene receives wake event
+    this.sys.events.on('wake', this.createMenu, this);
+
+    // the message describing the current action
+    this.message = new Message(this, this.battleScene.events);
+    this.add.existing(this.message);
+
+    this.createMenu();
+  },
+
+  createMenu() {
+    // map hero menu items to heroes
+    this.remapHeroes();
+    // map enemies menu items to enemies
+    this.remapEnemies();
+    // first move
+    this.battleScene.nextTurn();
+  },
+
+  onEnemy(index) {
+  // when the enemy is selected, we deselect all menus and send event with the enemy id
+    this.heroesMenu.deselect();
+    this.actionsMenu.deselect();
+    this.enemiesMenu.deselect();
+    this.currentMenu = null;
+    this.battleScene.receivePlayerSelection('attack', index);
+  },
+
+  onPlayerSelect(id) {
+    // when its player turn, we select the active hero item and the first action
+    // then we make actions menu active
+    this.heroesMenu.select(id);
+    this.actionsMenu.select(0);
+    this.currentMenu = this.actionsMenu;
+  },
+  // we have action selected and we make the enemies menu active
+  // the player needs to choose an enemy to attack
+  onSelectedAction() {
+    this.currentMenu = this.enemiesMenu;
+    this.enemiesMenu.select(0);
+  },
+
+  remapHeroes() {
+    const { heroes } = this.battleScene.heroes;
+    this.heroesMenu.remap(heroes);
+  },
+
+  remapEnemies() {
+    const { enemies } = this.battleScene.enemies;
+    this.enemiesMenu.remap(enemies);
+  },
+
+  onKeyInput(event) {
+    if (this.currentMenu && this.currentMenu.selected) {
+      if (event.code === 'ArrowUp') {
+        this.currentMenu.moveSelectionUp();
+      } else if (event.code === 'ArrowDown') {
+        this.currentMenu.moveSelectionDown();
+      } else if (event.code === 'ArrowRight' || event.code === 'Shift') {
+
+      } else if (event.code === 'Space' || event.code === 'ArrowLeft') {
+        this.currentMenu.confirm();
+      }
+    }
+  },
+});
+
+// the message class extends containter
+const Message = new Phaser.Class({
+
+  Extends: Phaser.GameObjects.Container,
+
+  initialize:
+  function Message(scene, events) {
+    Phaser.GameObjects.Container.call(this, scene, 160, 30);
+    const graphics = this.scene.add.graphics();
+    this.add(graphics);
+    graphics.lineStyle(1, 0xffffff, 0.8);
+    graphics.fillStyle(0x031f4c, 0.3);
+    graphics.strokeRect(-90, -15, 180, 30);
+    graphics.fillRect(-90, -15, 180, 30);
+    this.text = new Phaser.GameObjects.Text(scene, 0, 0, '', {
+      color: '#ffffff', align: 'center', fontSize: 13, wordWrap: { width: 170, useAdvancedWrap: true },
+    });
+    this.add(this.text);
+    this.text.setOrigin(0.5);
+    events.on('Message', this.showMessage, this);
+    this.visible = false;
+  },
+  showMessage(text) {
+    this.text.setText(text);
+    this.visible = true;
+    if (this.hideEvent) {
+      this.hideEvent.remove(false);
+      this.hideEvent = this.scene.time.addEvent({
+        delay: 2000, callback: this.hideMessage, callbackScope: this,
+      });
+    }
+  },
+
+  hideMessage() {
+    this.hideEvent = null;
+    this.visible = false;
+  },
+});
+
+export { BattleScene, UIScene };
