--- src/Scene/UiScene.js
@@ -1,154 +1,0 @@
-import Phaser from 'phaser';
-
-const UIScene = new Phaser.Class({
-
-  Extends: Phaser.Scene,
-
-  initialize:
-
-    function UIScene() {
-      Phaser.Scene.call(this, { key: 'UIScene' });
-    },
-
-  create() {
-    // draw some background for the menu
-    this.graphics = this.add.graphics();
-    this.graphics.lineStyle(1, 0xffffff);
-    this.graphics.fillStyle(0x031f4c, 1);
-    this.graphics.strokeRect(2, 150, 90, 100);
-    this.graphics.fillRect(2, 150, 90, 100);
-    this.graphics.strokeRect(95, 150, 90, 100);
-    this.graphics.fillRect(95, 150, 90, 100);
-    this.graphics.strokeRect(188, 150, 130, 100);
-    this.graphics.fillRect(188, 150, 130, 100);
-
-    // basic container to hold all menus
-    this.menus = this.add.container();
-
-    this.heroesMenu = new HeroesMenu(195, 153, this);
-    this.actionsMenu = new ActionsMenu(100, 153, this);
-    this.enemiesMenu = new EnemiesMenu(8, 153, this);
-
-    // the currently selected menu 
-    this.currentMenu = this.actionsMenu;
-
-    // add menus to the container
-    this.menus.add(this.heroesMenu);
-    this.menus.add(this.actionsMenu);
-    this.menus.add(this.enemiesMenu);
-
-    this.battleScene = this.scene.get('BattleScene');
-
-    // listen for keyboard events
-    this.input.keyboard.on('keydown', this.onKeyInput, this);   
-
-    // when its player cunit turn to move
-    this.battleScene.events.on('PlayerSelect', this.onPlayerSelect, this);
-
-    // when the action on the menu is selected
-    // for now we have only one action so we dont send and action id
-    this.events.on('SelectedAction', this.onSelectedAction, this);
-
-    // an enemy is selected
-    this.events.on('Enemy', this.onEnemy, this);
-
-    // when the scene receives wake event
-    this.sys.events.on('wake', this.createMenu, this);
-
-    // the message describing the current action
-    this.message = new Message(this, this.battleScene.events);
-    this.add.existing(this.message);
-
-    this.createMenu();
-  },
-  createMenu() {
-    // map hero menu items to heroes
-    this.remapHeroes();
-    // map enemies menu items to enemies
-    this.remapEnemies();
-    // first move
-    this.battleScene.nextTurn(); 
-  },
-  onEnemy(index) {
-  // when the enemy is selected, we deselect all menus and send event with the enemy id
-    this.heroesMenu.deselect();
-    this.actionsMenu.deselect();
-    this.enemiesMenu.deselect();
-    this.currentMenu = null;
-    this.battleScene.receivePlayerSelection('attack', index);   
-  },
-
-  onPlayerSelect(id) {
-    // when its player turn, we select the active hero item and the first action
-    // then we make actions menu active
-    this.heroesMenu.select(id);
-    this.actionsMenu.select(0);
-    this.currentMenu = this.actionsMenu;
-  },
-  // we have action selected and we make the enemies menu active
-  // the player needs to choose an enemy to attack
-  onSelectedAction() {
-    this.currentMenu = this.enemiesMenu;
-    this.enemiesMenu.select(0);
-  },
-  remapHeroes() {
-    const { heroes } = this.battleScene.heroes;
-    this.heroesMenu.remap(heroes);
-  },
-  remapEnemies() {
-    const { enemies } = this.battleScene.enemies;
-    this.enemiesMenu.remap(enemies);
-  },
-  onKeyInput(event) {
-    if (this.currentMenu && this.currentMenu.selected) {
-      if (event.code === 'ArrowUp') {
-        this.currentMenu.moveSelectionUp();
-      } else if (event.code === 'ArrowDown') {
-        this.currentMenu.moveSelectionDown();
-      } else if (event.code === 'ArrowRight' || event.code === 'Shift') {
-
-            } else if (event.code === 'Space' || event.code === 'ArrowLeft') {
-        this.currentMenu.confirm();
-      }
-    }
-  },
-});
-
-// the message class extends containter 
-const Message = new Phaser.Class({
-
-  Extends: Phaser.GameObjects.Container,
-
-  initialize:
-  function Message(scene, events) {
-    Phaser.GameObjects.Container.call(this, scene, 160, 30);
-    const graphics = this.scene.add.graphics();
-    this.add(graphics);
-    graphics.lineStyle(1, 0xffffff, 0.8);
-    graphics.fillStyle(0x031f4c, 0.3);
-    graphics.strokeRect(-90, -15, 180, 30);
-    graphics.fillRect(-90, -15, 180, 30);
-    this.text = new Phaser.GameObjects.Text(scene, 0, 0, '', { color: '#ffffff', align: 'center', fontSize: 13, wordWrap: { width: 170, useAdvancedWrap: true }});
-    this.add(this.text);
-    this.text.setOrigin(0.5);
-    events.on('Message', this.showMessage, this);
-    this.visible = false;
-  },
-  showMessage(text) {
-    this.text.setText(text);
-    this.visible = true;
-    if (this.hideEvent) {
-      this.hideEvent.remove(false);
-      this.hideEvent = this.scene.time.addEvent({
-        delay: 2000, callback: this.hideMessage, callbackScope: this,
-      });
-    }
-  },
-
-  hideMessage() {
-    this.hideEvent = null;
-    this.visible = false;
-  },
-});
-
-export default UIScene;
