+++ src/Scene/worldScene.js
@@ -26,9 +26,9 @@ const WorldScene = new Phaser.Class({
 
     this.anims.create({
       key: 'left',
-      frames: this.anims.generateFrameNumbers('player', { frames: [1, 7, 1, 13]}),
+      frames: this.anims.generateFrameNumbers('player', { frames: [1, 7, 1, 13] }),
       frameRate: 10,
-      repeat: -1
+      repeat: -1,
     });
 
     // animation with key 'right'
@@ -40,13 +40,13 @@ const WorldScene = new Phaser.Class({
     });
     this.anims.create({
       key: 'up',
-      frames: this.anims.generateFrameNumbers('player', { frames: [2, 8, 2, 14]}),
+      frames: this.anims.generateFrameNumbers('player', { frames: [2, 8, 2, 14] }),
       frameRate: 10,
       repeat: -1,
     });
     this.anims.create({
       key: 'down',
-      frames: this.anims.generateFrameNumbers('player', { frames: [ 0, 6, 0, 12 ] }),
+      frames: this.anims.generateFrameNumbers('player', { frames: [0, 6, 0, 12] }),
       frameRate: 10,
       repeat: -1,
     });
